refactor: rename normalizer example
