+++ src/showcase/app/public/public-examples/autocomplete-examples/BUILD.bazel
@@ -11,14 +11,14 @@ ng_ts_library(
         "autocomplete-examples.module.ts",
         "autocomplete-forms-example/autocomplete-forms-example.component.ts",
         "autocomplete-hint-example/autocomplete-hint-example.component.ts",
-        "autocomplete-normalizer-example/autocomplete-locale-normalizer-example.component.ts",
+        "autocomplete-locale-normalizer-example/autocomplete-locale-normalizer-example.component.ts",
         "autocomplete-option-group-example/autocomplete-option-group-example.component.ts",
         "autocomplete-reactive-forms-example/autocomplete-reactive-forms-example.component.ts",
     ],
     assets = [
         "autocomplete-forms-example/autocomplete-forms-example.component.html",
         "autocomplete-hint-example/autocomplete-hint-example.component.html",
-        "autocomplete-normalizer-example/autocomplete-locale-normalizer-example.component.html",
+        "autocomplete-locale-normalizer-example/autocomplete-locale-normalizer-example.component.html",
         "autocomplete-option-group-example/autocomplete-option-group-example.component.html",
         "autocomplete-reactive-forms-example/autocomplete-reactive-forms-example.component.html",
     ],
