+++ src/showcase/app/public/public-examples/autocomplete-examples/autocomplete-examples.module.ts
@@ -6,7 +6,7 @@ import { FieldModule } from '@sbb-esta/angular-public/field';
 
 import { AutocompleteFormsExampleComponent } from './autocomplete-forms-example/autocomplete-forms-example.component';
 import { AutocompleteHintExampleComponent } from './autocomplete-hint-example/autocomplete-hint-example.component';
-import { AutocompleteLocaleNormalizerExampleComponent } from './autocomplete-normalizer-example/autocomplete-locale-normalizer-example.component';
+import { AutocompleteLocaleNormalizerExampleComponent } from './autocomplete-locale-normalizer-example/autocomplete-locale-normalizer-example.component';
 import { AutocompleteOptionGroupExampleComponent } from './autocomplete-option-group-example/autocomplete-option-group-example.component';
 import { AutocompleteReactiveFormsExampleComponent } from './autocomplete-reactive-forms-example/autocomplete-reactive-forms-example.component';
